Make parser ignore all non-instruction characters
Introduced method `push_instruction` to keep track of the indices of the
instructions in the program so that loops can jump to the right places.
The method is also meant to keep track of instruction arguments when
this is added in the future.

Changed some tests and added tests for the loop instructions